Combinations of NLCD land cover classes into more general groups for Pennsylvania, USA.

Landcover within a 1km buffer of Pennsylvania BBS routes. Original cover classes combined into overarching groups. Original data from the National Land Cover Database (NLCD) https://catalog.data.gov/dataset/national-land-cover-database-nlcd-land-cover-collection The data were taken from the 2006 NLCD dataset.

BBS_PA_combo_cover_1km

Format

A data frame...

Route

BBS route number

BCR

Bird Conservation Region

State

State

route_name

Route name. Originally RT..NAME

route_length

length of route. Originally RT.LENG

open_H2O

NLCD.11

developed

NLCD.21,22,23,24

forest

NLCD.41,42,43

grassland

NLCD.71

pasture

NLCD.81

crops

NLCD.82

wetlands

NLCD.90,95

cov_sum

total landcover accross all variables, not just those in the combined class. Orginally SUM

Source

https://figshare.com/articles/Data_Paper_Data_Paper/3554424

References

Small et al. 2016. Data Paper: Local landscape composition and configuration around North American Breeding Bird Survey routes. See also: http://esapubs.org/archive/ecol/E093/215/