Land cover buffers from the 1 km surrounding Breeding Bird Survey Routes (BBS) in Pennsylvania

Landcover within a 1km buffer of Pennsylvania BBS routes. Original data from the National Land Cover Database (NLCD) https://catalog.data.gov/dataset/national-land-cover-database-nlcd-land-cover-collection The data were taken from the 2006 NLCD dataset.

BBS_PA_landcover_1km

Format

A data frame...

Route

BBS route number

BCR

Bird conservation region

State

State

RT..NAME

Route name

RT..LENG.

length of route

NLCD.0

amount of habitat in category 0 of the National Land Cover Database

NLCD.11

Open water

NLCD.12

NLCD.12

NLCD.21

Developed Open Space; town and city parks, recreational fields, large-lot neighborhoods

NLCD.22

Developed Low Intensity; 1 of 4 developed cover classes

NLCD.23

Developed Medium Intensity; 1 of 4 developed cover classes

NLCD.24

Developed High Intensity; 1 of 4 developed cover classes

NLCD.31

Barren land

NLCD.41

Deciduous Forest; 1 of 3 forest landcover classes

NLCD.42

Evergreen Forest; 1 of 3 forest landcover classes

NLCD.43

Mixed Forest; 1 of 3forest landcover classes

NLCD.52

Shrub or scrub

NLCD.71

Grassland or Herbaceous

NLCD.81

Pasture or Hay

NLCD.82

Cultivated Crops

NLCD.90

Woody wetland

NLCD.95

Herbaceous wetland

SUM

total landcover accross all variables

Source

https://figshare.com/articles/Data_Paper_Data_Paper/3554424

References

Small et al. 2016. Data Paper: Local landscape composition and configuration around North American Breeding Bird Survey routes. See also: http://esapubs.org/archive/ecol/E093/215/

Veech et al. 2012. Representativeness f land cover composition along routes of the North American Breeding Bird Survey. Aug 129:259-267.