Impact of brook trout on survival of Chinook salmon in Idaho.

A dataset ...

brook_trout_ABD

Format

A data frame with 12 rows and 3 columns

brook.trout.PRES.ABS

Whether brook trout are present or absent

salmon.released

Number of PIT-tagged salmon released

salmon.surv

Number of tagged salmon that survived and were detected downstream

Source

http://rspb.royalsocietypublishing.org/content/269/1501/1663.short

References

Levin et al. 2002. Non–indigenous brook trout and the demise of Pacific salmon: a forgotten threat? PRSB 269. DOI: 10.1098/rspb.2002.2063

Whitlock & Schulter. 2nd ed. Analysis of Biological Data.

Examples

## t-test on survival rate (ignores years)

### Calcualte survival
brook_trout_ABD$surv <- brook_trout_ABD$salmon.surv/brook_trout_ABD$salmon.released

### t-test
t.test(surv ~ brook.trout.PRES.ABS, data = brook_trout_ABD)

#> 
#> 	Welch Two Sample t-test
#> 
#> data:  surv by brook.trout.PRES.ABS
#> t = 0.93521, df = 5.8196, p-value = 0.3868
#> alternative hypothesis: true difference in means between group absent and group present is not equal to 0
#> 95 percent confidence interval:
#>  -0.06739746  0.14977890
#> sample estimates:
#>  mean in group absent mean in group present 
#>             0.2352653             0.1940746 
#>