Combine multiple landcover classes into a single combined group

This function takes several landcover classes, such as all forest landcover classes and totals them up. It is simply wrapper for apply(...sum) that

combine_NLCD_classes(dat, numeric_cover_codes)

Arguments

dat
dataframe of NLCD landcover values

numeric_cover_codes
numeric codes for landcover classes. Do not include the "NCLE" prefix.