Function to convert a FASTA file stored as an object into a vector

Function to convert a FASTA file stored as an object into a vector

fasta_cleaner(fasta_object, parse = TRUE)

Arguments

fasta_object
R object with FASTA character data to clean

parse
Parse data to character vector?

Examples

dengueseq_fasta <- rentrez::entrez_fetch(db = "nucleotide", id = "NC_001477", rettype = "fasta")
fasta_cleaner(dengueseq_fasta, parse = TRUE)

#>     [1] "A" "G" "T" "T" "G" "T" "T" "A" "G" "T" "C" "T" "A" "C" "G" "T" "G" "G"
#>    [19] "A" "C" "C" "G" "A" "C" "A" "A" "G" "A" "A" "C" "A" "G" "T" "T" "T" "C"
#>    [37] "G" "A" "A" "T" "C" "G" "G" "A" "A" "G" "C" "T" "T" "G" "C" "T" "T" "A"
#>    [55] "A" "C" "G" "T" "A" "G" "T" "T" "C" "T" "A" "A" "C" "A" "G" "T" "T" "T"
#>    [73] "T" "T" "T" "A" "T" "T" "A" "G" "A" "G" "A" "G" "C" "A" "G" "A" "T" "C"
#>    [91] "T" "C" "T" "G" "A" "T" "G" "A" "A" "C" "A" "A" "C" "C" "A" "A" "C" "G"
#>   [109] "G" "A" "A" "A" "A" "A" "G" "A" "C" "G" "G" "G" "T" "C" "G" "A" "C" "C"
#>   [127] "G" "T" "C" "T" "T" "T" "C" "A" "A" "T" "A" "T" "G" "C" "T" "G" "A" "A"
#>   [145] "A" "C" "G" "C" "G" "C" "G" "A" "G" "A" "A" "A" "C" "C" "G" "C" "G" "T"
#>   [163] "G" "T" "C" "A" "A" "C" "T" "G" "T" "T" "T" "C" "A" "C" "A" "G" "T" "T"
#>   [181] "G" "G" "C" "G" "A" "A" "G" "A" "G" "A" "T" "T" "C" "T" "C" "A" "A" "A"
#>   [199] "A" "G" "G" "A" "T" "T" "G" "C" "T" "T" "T" "C" "A" "G" "G" "C" "C" "A"
#>   [217] "A" "G" "G" "A" "C" "C" "C" "A" "T" "G" "A" "A" "A" "T" "T" "G" "G" "T"
#>   [235] "G" "A" "T" "G" "G" "C" "T" "T" "T" "T" "A" "T" "A" "G" "C" "A" "T" "T"
#>   [253] "C" "C" "T" "A" "A" "G" "A" "T" "T" "T" "C" "T" "A" "G" "C" "C" "A" "T"
#>   [271] "A" "C" "C" "T" "C" "C" "A" "A" "C" "A" "G" "C" "A" "G" "G" "A" "A" "T"
#>   [289] "T" "T" "T" "G" "G" "C" "T" "A" "G" "A" "T" "G" "G" "G" "G" "C" "T" "C"
#>   [307] "A" "T" "T" "C" "A" "A" "G" "A" "A" "G" "A" "A" "T" "G" "G" "A" "G" "C"
#>   [325] "G" "A" "T" "C" "A" "A" "A" "G" "T" "G" "T" "T" "A" "C" "G" "G" "G" "G"
#>   [343] "T" "T" "T" "C" "A" "A" "G" "A" "A" "A" "G" "A" "A" "A" "T" "C" "T" "C"
#>   [361] "A" "A" "A" "C" "A" "T" "G" "T" "T" "G" "A" "A" "C" "A" "T" "A" "A" "T"
#>   [379] "G" "A" "A" "C" "A" "G" "G" "A" "G" "G" "A" "A" "A" "A" "G" "A" "T" "C"
#>   [397] "T" "G" "T" "G" "A" "C" "C" "A" "T" "G" "C" "T" "C" "C" "T" "C" "A" "T"
#>   [415] "G" "C" "T" "G" "C" "T" "G" "C" "C" "C" "A" "C" "A" "G" "C" "C" "C" "T"
#>   [433] "G" "G" "C" "G" "T" "T" "C" "C" "A" "T" "C" "T" "G" "A" "C" "C" "A" "C"
#>   [451] "C" "C" "G" "A" "G" "G" "G" "G" "G" "A" "G" "A" "G" "C" "C" "G" "C" "A"
#>   [469] "C" "A" "T" "G" "A" "T" "A" "G" "T" "T" "A" "G" "C" "A" "A" "G" "C" "A"
#>   [487] "G" "G" "A" "A" "A" "G" "A" "G" "G" "A" "A" "A" "A" "T" "C" "A" "C" "T"
#>   [505] "T" "T" "T" "G" "T" "T" "T" "A" "A" "G" "A" "C" "C" "T" "C" "T" "G" "C"
#>   [523] "A" "G" "G" "T" "G" "T" "C" "A" "A" "C" "A" "T" "G" "T" "G" "C" "A" "C"
#>   [541] "C" "C" "T" "T" "A" "T" "T" "G" "C" "A" "A" "T" "G" "G" "A" "T" "T" "T"
#>   [559] "G" "G" "G" "A" "G" "A" "G" "T" "T" "A" "T" "G" "T" "G" "A" "G" "G" "A"
#>   [577] "C" "A" "C" "A" "A" "T" "G" "A" "C" "C" "T" "A" "C" "A" "A" "A" "T" "G"
#>   [595] "C" "C" "C" "C" "C" "G" "G" "A" "T" "C" "A" "C" "T" "G" "A" "G" "A" "C"
#>   [613] "G" "G" "A" "A" "C" "C" "A" "G" "A" "T" "G" "A" "C" "G" "T" "T" "G" "A"
#>   [631] "C" "T" "G" "T" "T" "G" "G" "T" "G" "C" "A" "A" "T" "G" "C" "C" "A" "C"
#>   [649] "G" "G" "A" "G" "A" "C" "A" "T" "G" "G" "G" "T" "G" "A" "C" "C" "T" "A"
#>   [667] "T" "G" "G" "A" "A" "C" "A" "T" "G" "T" "T" "C" "T" "C" "A" "A" "A" "C"
#>   [685] "T" "G" "G" "T" "G" "A" "A" "C" "A" "C" "C" "G" "A" "C" "G" "A" "G" "A"
#>   [703] "C" "A" "A" "A" "C" "G" "T" "T" "C" "C" "G" "T" "C" "G" "C" "A" "C" "T"
#>   [721] "G" "G" "C" "A" "C" "C" "A" "C" "A" "C" "G" "T" "A" "G" "G" "G" "C" "T"
#>   [739] "T" "G" "G" "T" "C" "T" "A" "G" "A" "A" "A" "C" "A" "A" "G" "A" "A" "C"
#>   [757] "C" "G" "A" "A" "A" "C" "G" "T" "G" "G" "A" "T" "G" "T" "C" "C" "T" "C"
#>   [775] "T" "G" "A" "A" "G" "G" "C" "G" "C" "T" "T" "G" "G" "A" "A" "A" "C" "A"
#>   [793] "A" "A" "T" "A" "C" "A" "A" "A" "A" "A" "G" "T" "G" "G" "A" "G" "A" "C"
#>   [811] "C" "T" "G" "G" "G" "C" "T" "C" "T" "G" "A" "G" "A" "C" "A" "C" "C" "C"
#>   [829] "A" "G" "G" "A" "T" "T" "C" "A" "C" "G" "G" "T" "G" "A" "T" "A" "G" "C"
#>   [847] "C" "C" "T" "T" "T" "T" "T" "C" "T" "A" "G" "C" "A" "C" "A" "T" "G" "C"
#>   [865] "C" "A" "T" "A" "G" "G" "A" "A" "C" "A" "T" "C" "C" "A" "T" "C" "A" "C"
#>   [883] "C" "C" "A" "G" "A" "A" "A" "G" "G" "G" "A" "T" "C" "A" "T" "T" "T" "T"
#>   [901] "T" "A" "T" "T" "T" "T" "G" "C" "T" "G" "A" "T" "G" "C" "T" "G" "G" "T"
#>   [919] "A" "A" "C" "T" "C" "C" "A" "T" "C" "C" "A" "T" "G" "G" "C" "C" "A" "T"
#>   [937] "G" "C" "G" "G" "T" "G" "C" "G" "T" "G" "G" "G" "A" "A" "T" "A" "G" "G"
#>   [955] "C" "A" "A" "C" "A" "G" "A" "G" "A" "C" "T" "T" "C" "G" "T" "G" "G" "A"
#>   [973] "A" "G" "G" "A" "C" "T" "G" "T" "C" "A" "G" "G" "A" "G" "C" "T" "A" "C"
#>   [991] "G" "T" "G" "G" "G" "T" "G" "G" "A" "T" "G" "T" "G" "G" "T" "A" "C" "T"
#>  [1009] "G" "G" "A" "G" "C" "A" "T" "G" "G" "A" "A" "G" "T" "T" "G" "C" "G" "T"
#>  [1027] "C" "A" "C" "T" "A" "C" "C" "A" "T" "G" "G" "C" "A" "A" "A" "A" "G" "A"
#>  [1045] "C" "A" "A" "A" "C" "C" "A" "A" "C" "A" "C" "T" "G" "G" "A" "C" "A" "T"
#>  [1063] "T" "G" "A" "A" "C" "T" "C" "T" "T" "G" "A" "A" "G" "A" "C" "G" "G" "A"
#>  [1081] "G" "G" "T" "C" "A" "C" "A" "A" "A" "C" "C" "C" "T" "G" "C" "C" "G" "T"
#>  [1099] "C" "C" "T" "G" "C" "G" "C" "A" "A" "A" "C" "T" "G" "T" "G" "C" "A" "T"
#>  [1117] "T" "G" "A" "A" "G" "C" "T" "A" "A" "A" "A" "T" "A" "T" "C" "A" "A" "A"
#>  [1135] "C" "A" "C" "C" "A" "C" "C" "A" "C" "C" "G" "A" "T" "T" "C" "G" "A" "G"
#>  [1153] "A" "T" "G" "T" "C" "C" "A" "A" "C" "A" "C" "A" "A" "G" "G" "A" "G" "A"
#>  [1171] "A" "G" "C" "C" "A" "C" "G" "C" "T" "G" "G" "T" "G" "G" "A" "A" "G" "A"
#>  [1189] "A" "C" "A" "G" "G" "A" "C" "A" "C" "G" "A" "A" "C" "T" "T" "T" "G" "T"
#>  [1207] "G" "T" "G" "T" "C" "G" "A" "C" "G" "A" "A" "C" "G" "T" "T" "C" "G" "T"
#>  [1225] "G" "G" "A" "C" "A" "G" "A" "G" "G" "C" "T" "G" "G" "G" "G" "C" "A" "A"
#>  [1243] "T" "G" "G" "T" "T" "G" "T" "G" "G" "G" "C" "T" "A" "T" "T" "C" "G" "G"
#>  [1261] "A" "A" "A" "A" "G" "G" "T" "A" "G" "C" "T" "T" "A" "A" "T" "A" "A" "C"
#>  [1279] "G" "T" "G" "T" "G" "C" "T" "A" "A" "G" "T" "T" "T" "A" "A" "G" "T" "G"
#>  [1297] "T" "G" "T" "G" "A" "C" "A" "A" "A" "A" "C" "T" "G" "G" "A" "A" "G" "G"
#>  [1315] "A" "A" "A" "G" "A" "T" "A" "G" "T" "C" "C" "A" "A" "T" "A" "T" "G" "A"
#>  [1333] "A" "A" "A" "C" "T" "T" "A" "A" "A" "A" "T" "A" "T" "T" "C" "A" "G" "T"
#>  [1351] "G" "A" "T" "A" "G" "T" "C" "A" "C" "C" "G" "T" "A" "C" "A" "C" "A" "C"
#>  [1369] "T" "G" "G" "A" "G" "A" "C" "C" "A" "G" "C" "A" "C" "C" "A" "A" "G" "T"
#>  [1387] "T" "G" "G" "A" "A" "A" "T" "G" "A" "G" "A" "C" "C" "A" "C" "A" "G" "A"
#>  [1405] "A" "C" "A" "T" "G" "G" "A" "A" "C" "A" "A" "C" "T" "G" "C" "A" "A" "C"
#>  [1423] "C" "A" "T" "A" "A" "C" "A" "C" "C" "T" "C" "A" "A" "G" "C" "T" "C" "C"
#>  [1441] "C" "A" "C" "G" "T" "C" "G" "G" "A" "A" "A" "T" "A" "C" "A" "G" "C" "T"
#>  [1459] "G" "A" "C" "A" "G" "A" "C" "T" "A" "C" "G" "G" "A" "G" "C" "T" "C" "T"
#>  [1477] "A" "A" "C" "A" "T" "T" "G" "G" "A" "T" "T" "G" "T" "T" "C" "A" "C" "C"
#>  [1495] "T" "A" "G" "A" "A" "C" "A" "G" "G" "G" "C" "T" "A" "G" "A" "C" "T" "T"
#>  [1513] "T" "A" "A" "T" "G" "A" "G" "A" "T" "G" "G" "T" "G" "T" "T" "G" "T" "T"
#>  [1531] "G" "A" "C" "A" "A" "T" "G" "A" "A" "A" "A" "A" "A" "A" "A" "A" "T" "C"
#>  [1549] "A" "T" "G" "G" "C" "T" "C" "G" "T" "C" "C" "A" "C" "A" "A" "A" "C" "A"
#>  [1567] "A" "T" "G" "G" "T" "T" "T" "C" "T" "A" "G" "A" "C" "T" "T" "A" "C" "C"
#>  [1585] "A" "C" "T" "G" "C" "C" "T" "T" "G" "G" "A" "C" "C" "T" "C" "G" "G" "G"
#>  [1603] "G" "G" "C" "T" "T" "C" "A" "A" "C" "A" "T" "C" "C" "C" "A" "A" "G" "A"
#>  [1621] "G" "A" "C" "T" "T" "G" "G" "A" "A" "T" "A" "G" "A" "C" "A" "A" "G" "A"
#>  [1639] "C" "T" "T" "G" "C" "T" "G" "G" "T" "C" "A" "C" "A" "T" "T" "T" "A" "A"
#>  [1657] "G" "A" "C" "A" "G" "C" "T" "C" "A" "T" "G" "C" "A" "A" "A" "A" "A" "A"
#>  [1675] "G" "C" "A" "G" "G" "A" "A" "G" "T" "A" "G" "T" "C" "G" "T" "A" "C" "T"
#>  [1693] "A" "G" "G" "A" "T" "C" "A" "C" "A" "A" "G" "A" "A" "G" "G" "A" "G" "C"
#>  [1711] "A" "A" "T" "G" "C" "A" "C" "A" "C" "T" "G" "C" "G" "T" "T" "G" "A" "C"
#>  [1729] "T" "G" "G" "A" "G" "C" "G" "A" "C" "A" "G" "A" "A" "A" "T" "C" "C" "A"
#>  [1747] "A" "A" "C" "G" "T" "C" "T" "G" "G" "A" "A" "C" "G" "A" "C" "A" "A" "C"
#>  [1765] "A" "A" "T" "T" "T" "T" "T" "G" "C" "A" "G" "G" "A" "C" "A" "C" "C" "T"
#>  [1783] "G" "A" "A" "A" "T" "G" "C" "A" "G" "A" "T" "T" "A" "A" "A" "A" "A" "T"
#>  [1801] "G" "G" "A" "T" "A" "A" "A" "C" "T" "G" "A" "T" "T" "T" "T" "A" "A" "A"
#>  [1819] "A" "G" "G" "G" "A" "T" "G" "T" "C" "A" "T" "A" "T" "G" "T" "A" "A" "T"
#>  [1837] "G" "T" "G" "C" "A" "C" "A" "G" "G" "G" "T" "C" "A" "T" "T" "C" "A" "A"
#>  [1855] "G" "T" "T" "A" "G" "A" "G" "A" "A" "G" "G" "A" "A" "G" "T" "G" "G" "C"
#>  [1873] "T" "G" "A" "G" "A" "C" "C" "C" "A" "G" "C" "A" "T" "G" "G" "A" "A" "C"
#>  [1891] "T" "G" "T" "T" "C" "T" "A" "G" "T" "G" "C" "A" "G" "G" "T" "T" "A" "A"
#>  [1909] "A" "T" "A" "C" "G" "A" "A" "G" "G" "A" "A" "C" "A" "G" "A" "T" "G" "C"
#>  [1927] "A" "C" "C" "A" "T" "G" "C" "A" "A" "G" "A" "T" "C" "C" "C" "C" "T" "T"
#>  [1945] "C" "T" "C" "G" "T" "C" "C" "C" "A" "A" "G" "A" "T" "G" "A" "G" "A" "A"
#>  [1963] "G" "G" "G" "A" "G" "T" "A" "A" "C" "C" "C" "A" "G" "A" "A" "T" "G" "G"
#>  [1981] "G" "A" "G" "A" "T" "T" "G" "A" "T" "A" "A" "C" "A" "G" "C" "C" "A" "A"
#>  [1999] "C" "C" "C" "C" "A" "T" "A" "G" "T" "C" "A" "C" "T" "G" "A" "C" "A" "A"
#>  [2017] "A" "G" "A" "A" "A" "A" "A" "C" "C" "A" "G" "T" "C" "A" "A" "C" "A" "T"
#>  [2035] "T" "G" "A" "A" "G" "C" "G" "G" "A" "G" "C" "C" "A" "C" "C" "T" "T" "T"
#>  [2053] "T" "G" "G" "T" "G" "A" "G" "A" "G" "C" "T" "A" "C" "A" "T" "T" "G" "T"
#>  [2071] "G" "G" "T" "A" "G" "G" "A" "G" "C" "A" "G" "G" "T" "G" "A" "A" "A" "A"
#>  [2089] "A" "G" "C" "T" "T" "T" "G" "A" "A" "A" "C" "T" "A" "A" "G" "C" "T" "G"
#>  [2107] "G" "T" "T" "C" "A" "A" "G" "A" "A" "G" "G" "G" "A" "A" "G" "C" "A" "G"
#>  [2125] "T" "A" "T" "A" "G" "G" "G" "A" "A" "A" "A" "T" "G" "T" "T" "T" "G" "A"
#>  [2143] "A" "G" "C" "A" "A" "C" "T" "G" "C" "C" "C" "G" "T" "G" "G" "A" "G" "C"
#>  [2161] "A" "C" "G" "A" "A" "G" "G" "A" "T" "G" "G" "C" "C" "A" "T" "C" "C" "T"
#>  [2179] "G" "G" "G" "A" "G" "A" "C" "A" "C" "T" "G" "C" "A" "T" "G" "G" "G" "A"
#>  [2197] "C" "T" "T" "C" "G" "G" "T" "T" "C" "T" "A" "T" "A" "G" "G" "A" "G" "G"
#>  [2215] "G" "G" "T" "G" "T" "T" "C" "A" "C" "G" "T" "C" "T" "G" "T" "G" "G" "G"
#>  [2233] "A" "A" "A" "A" "C" "T" "G" "A" "T" "A" "C" "A" "C" "C" "A" "G" "A" "T"
#>  [2251] "T" "T" "T" "T" "G" "G" "G" "A" "C" "T" "G" "C" "G" "T" "A" "T" "G" "G"
#>  [2269] "A" "G" "T" "T" "T" "T" "G" "T" "T" "C" "A" "G" "C" "G" "G" "T" "G" "T"
#>  [2287] "T" "T" "C" "T" "T" "G" "G" "A" "C" "C" "A" "T" "G" "A" "A" "G" "A" "T"
#>  [2305] "A" "G" "G" "A" "A" "T" "A" "G" "G" "G" "A" "T" "T" "C" "T" "G" "C" "T"
#>  [2323] "G" "A" "C" "A" "T" "G" "G" "C" "T" "A" "G" "G" "A" "T" "T" "A" "A" "A"
#>  [2341] "C" "T" "C" "A" "A" "G" "G" "A" "G" "C" "A" "C" "G" "T" "C" "C" "C" "T"
#>  [2359] "T" "T" "C" "A" "A" "T" "G" "A" "C" "G" "T" "G" "T" "A" "T" "C" "G" "C"
#>  [2377] "A" "G" "T" "T" "G" "G" "C" "A" "T" "G" "G" "T" "C" "A" "C" "A" "C" "T"
#>  [2395] "G" "T" "A" "C" "C" "T" "A" "G" "G" "A" "G" "T" "C" "A" "T" "G" "G" "T"
#>  [2413] "T" "C" "A" "G" "G" "C" "G" "G" "A" "C" "T" "C" "G" "G" "G" "A" "T" "G"
#>  [2431] "T" "G" "T" "A" "A" "T" "C" "A" "A" "C" "T" "G" "G" "A" "A" "A" "G" "G"
#>  [2449] "C" "A" "G" "A" "G" "A" "A" "C" "T" "C" "A" "A" "A" "T" "G" "T" "G" "G"
#>  [2467] "A" "A" "G" "C" "G" "G" "C" "A" "T" "T" "T" "T" "T" "G" "T" "C" "A" "C"
#>  [2485] "C" "A" "A" "T" "G" "A" "A" "G" "T" "C" "C" "A" "C" "A" "C" "C" "T" "G"
#>  [2503] "G" "A" "C" "A" "G" "A" "G" "C" "A" "A" "T" "A" "T" "A" "A" "A" "T" "T"
#>  [2521] "C" "C" "A" "G" "G" "C" "C" "G" "A" "C" "T" "C" "C" "C" "C" "T" "A" "A"
#>  [2539] "G" "A" "G" "A" "C" "T" "A" "T" "C" "A" "G" "C" "G" "G" "C" "C" "A" "T"
#>  [2557] "T" "G" "G" "G" "A" "A" "G" "G" "C" "A" "T" "G" "G" "G" "A" "G" "G" "A"
#>  [2575] "G" "G" "G" "T" "G" "T" "G" "T" "G" "T" "G" "G" "A" "A" "T" "T" "C" "G"
#>  [2593] "A" "T" "C" "A" "G" "C" "C" "A" "C" "T" "C" "G" "T" "C" "T" "C" "G" "A"
#>  [2611] "G" "A" "A" "C" "A" "T" "C" "A" "T" "G" "T" "G" "G" "A" "A" "G" "C" "A"
#>  [2629] "A" "A" "T" "A" "T" "C" "A" "A" "A" "T" "G" "A" "A" "T" "T" "A" "A" "A"
#>  [2647] "C" "C" "A" "C" "A" "T" "C" "T" "T" "A" "C" "T" "T" "G" "A" "A" "A" "A"
#>  [2665] "T" "G" "A" "C" "A" "T" "G" "A" "A" "A" "T" "T" "T" "A" "C" "A" "G" "T"
#>  [2683] "G" "G" "T" "C" "G" "T" "A" "G" "G" "A" "G" "A" "C" "G" "T" "T" "A" "G"
#>  [2701] "T" "G" "G" "A" "A" "T" "C" "T" "T" "G" "G" "C" "C" "C" "A" "A" "G" "G"
#>  [2719] "A" "A" "A" "G" "A" "A" "A" "A" "T" "G" "A" "T" "T" "A" "G" "G" "C" "C"
#>  [2737] "A" "C" "A" "A" "C" "C" "C" "A" "T" "G" "G" "A" "A" "C" "A" "C" "A" "A"
#>  [2755] "A" "T" "A" "C" "T" "C" "G" "T" "G" "G" "A" "A" "A" "A" "G" "C" "T" "G"
#>  [2773] "G" "G" "G" "A" "A" "A" "A" "G" "C" "C" "A" "A" "A" "A" "T" "C" "A" "T"
#>  [2791] "A" "G" "G" "A" "G" "C" "A" "G" "A" "T" "G" "T" "A" "C" "A" "G" "A" "A"
#>  [2809] "T" "A" "C" "C" "A" "C" "C" "T" "T" "C" "A" "T" "C" "A" "T" "C" "G" "A"
#>  [2827] "C" "G" "G" "C" "C" "C" "A" "A" "A" "C" "A" "C" "C" "C" "C" "A" "G" "A"
#>  [2845] "A" "T" "G" "C" "C" "C" "T" "G" "A" "T" "A" "A" "C" "C" "A" "A" "A" "G"
#>  [2863] "A" "G" "C" "A" "T" "G" "G" "A" "A" "C" "A" "T" "T" "T" "G" "G" "G" "A"
#>  [2881] "A" "G" "T" "T" "G" "A" "A" "G" "A" "C" "T" "A" "T" "G" "G" "A" "T" "T"
#>  [2899] "T" "G" "G" "A" "A" "T" "T" "T" "T" "C" "A" "C" "G" "A" "C" "A" "A" "A"
#>  [2917] "C" "A" "T" "A" "T" "G" "G" "T" "T" "G" "A" "A" "A" "T" "T" "G" "C" "G"
#>  [2935] "T" "G" "A" "C" "T" "C" "C" "T" "A" "C" "A" "C" "T" "C" "A" "A" "G" "T"
#>  [2953] "G" "T" "G" "T" "G" "A" "C" "C" "A" "C" "C" "G" "G" "C" "T" "A" "A" "T"
#>  [2971] "G" "T" "C" "A" "G" "C" "T" "G" "C" "C" "A" "T" "C" "A" "A" "G" "G" "A"
#>  [2989] "T" "A" "G" "C" "A" "A" "A" "G" "C" "A" "G" "T" "C" "C" "A" "T" "G" "C"
#>  [3007] "T" "G" "A" "C" "A" "T" "G" "G" "G" "G" "T" "A" "C" "T" "G" "G" "A" "T"
#>  [3025] "A" "G" "A" "A" "A" "G" "T" "G" "A" "A" "A" "A" "G" "A" "A" "C" "G" "A"
#>  [3043] "G" "A" "C" "T" "T" "G" "G" "A" "A" "G" "T" "T" "G" "G" "C" "A" "A" "G"
#>  [3061] "A" "G" "C" "C" "T" "C" "C" "T" "T" "C" "A" "T" "A" "G" "A" "A" "G" "T"
#>  [3079] "T" "A" "A" "G" "A" "C" "A" "T" "G" "C" "A" "T" "C" "T" "G" "G" "C" "C"
#>  [3097] "A" "A" "A" "A" "T" "C" "C" "C" "A" "C" "A" "C" "T" "C" "T" "A" "T" "G"
#>  [3115] "G" "A" "G" "C" "A" "A" "T" "G" "G" "A" "G" "T" "C" "C" "T" "G" "G" "A"
#>  [3133] "A" "A" "G" "T" "G" "A" "G" "A" "T" "G" "A" "T" "A" "A" "T" "C" "C" "C"
#>  [3151] "A" "A" "A" "G" "A" "T" "A" "T" "A" "T" "G" "G" "A" "G" "G" "A" "C" "C"
#>  [3169] "A" "A" "T" "A" "T" "C" "T" "C" "A" "G" "C" "A" "C" "A" "A" "C" "T" "A"
#>  [3187] "C" "A" "G" "A" "C" "C" "A" "G" "G" "A" "T" "A" "T" "T" "T" "C" "A" "C"
#>  [3205] "A" "C" "A" "A" "A" "C" "A" "G" "C" "A" "G" "G" "G" "C" "C" "G" "T" "G"
#>  [3223] "G" "C" "A" "C" "T" "T" "G" "G" "G" "C" "A" "A" "G" "T" "T" "A" "G" "A"
#>  [3241] "A" "C" "T" "A" "G" "A" "T" "T" "T" "T" "G" "A" "T" "T" "T" "A" "T" "G"
#>  [3259] "T" "G" "A" "A" "G" "G" "T" "A" "C" "C" "A" "C" "T" "G" "T" "T" "G" "T"
#>  [3277] "T" "G" "T" "G" "G" "A" "T" "G" "A" "A" "C" "A" "T" "T" "G" "T" "G" "G"
#>  [3295] "A" "A" "A" "T" "C" "G" "A" "G" "G" "A" "C" "C" "A" "T" "C" "T" "C" "T"
#>  [3313] "T" "A" "G" "A" "A" "C" "C" "A" "C" "A" "A" "C" "A" "G" "T" "C" "A" "C"
#>  [3331] "A" "G" "G" "A" "A" "A" "G" "A" "C" "A" "A" "T" "C" "C" "A" "T" "G" "A"
#>  [3349] "A" "T" "G" "G" "T" "G" "C" "T" "G" "T" "A" "G" "A" "T" "C" "T" "T" "G"
#>  [3367] "C" "A" "C" "G" "T" "T" "A" "C" "C" "C" "C" "C" "C" "C" "T" "A" "C" "G"
#>  [3385] "T" "T" "T" "C" "A" "A" "A" "G" "G" "A" "G" "A" "A" "G" "A" "C" "G" "G"
#>  [3403] "G" "T" "G" "C" "T" "G" "G" "T" "A" "C" "G" "G" "C" "A" "T" "G" "G" "A"
#>  [3421] "A" "A" "T" "C" "A" "G" "A" "C" "C" "A" "G" "T" "C" "A" "A" "G" "G" "A"
#>  [3439] "G" "A" "A" "G" "G" "A" "A" "G" "A" "G" "A" "A" "C" "C" "T" "A" "G" "T"
#>  [3457] "T" "A" "A" "G" "T" "C" "A" "A" "T" "G" "G" "T" "C" "T" "C" "T" "G" "C"
#>  [3475] "A" "G" "G" "G" "T" "C" "A" "G" "G" "A" "G" "A" "A" "G" "T" "G" "G" "A"
#>  [3493] "C" "A" "G" "T" "T" "T" "T" "T" "C" "A" "C" "T" "A" "G" "G" "A" "C" "T"
#>  [3511] "G" "C" "T" "A" "T" "G" "C" "A" "T" "A" "T" "C" "A" "A" "T" "A" "A" "T"
#>  [3529] "G" "A" "T" "C" "G" "A" "A" "G" "A" "G" "G" "T" "A" "A" "T" "G" "A" "G"
#>  [3547] "A" "T" "C" "C" "A" "G" "A" "T" "G" "G" "A" "G" "C" "A" "G" "A" "A" "A"
#>  [3565] "A" "A" "T" "G" "C" "T" "G" "A" "T" "G" "A" "C" "T" "G" "G" "A" "A" "C"
#>  [3583] "A" "T" "T" "G" "G" "C" "T" "G" "T" "G" "T" "T" "C" "C" "T" "C" "C" "T"
#>  [3601] "T" "C" "T" "C" "A" "C" "A" "A" "T" "G" "G" "G" "A" "C" "A" "A" "T" "T"
#>  [3619] "G" "A" "C" "A" "T" "G" "G" "A" "A" "T" "G" "A" "T" "C" "T" "G" "A" "T"
#>  [3637] "C" "A" "G" "G" "C" "T" "A" "T" "G" "T" "A" "T" "C" "A" "T" "G" "G" "T"
#>  [3655] "T" "G" "G" "A" "G" "C" "C" "A" "A" "C" "G" "C" "T" "T" "C" "A" "G" "A"
#>  [3673] "C" "A" "A" "G" "A" "T" "G" "G" "G" "G" "A" "T" "G" "G" "G" "A" "A" "C"
#>  [3691] "A" "A" "C" "G" "T" "A" "C" "C" "T" "A" "G" "C" "T" "T" "T" "G" "A" "T"
#>  [3709] "G" "G" "C" "C" "A" "C" "T" "T" "T" "C" "A" "G" "A" "A" "T" "G" "A" "G"
#>  [3727] "A" "C" "C" "A" "A" "T" "G" "T" "T" "C" "G" "C" "A" "G" "T" "C" "G" "G"
#>  [3745] "G" "C" "T" "A" "C" "T" "G" "T" "T" "T" "C" "G" "C" "A" "G" "A" "T" "T"
#>  [3763] "A" "A" "C" "A" "T" "C" "T" "A" "G" "A" "G" "A" "A" "G" "T" "T" "C" "T"
#>  [3781] "T" "C" "T" "T" "C" "T" "T" "A" "C" "A" "G" "T" "T" "G" "G" "A" "T" "T"
#>  [3799] "G" "A" "G" "T" "C" "T" "G" "G" "T" "G" "G" "C" "A" "T" "C" "T" "G" "T"
#>  [3817] "A" "G" "A" "A" "C" "T" "A" "C" "C" "A" "A" "A" "T" "T" "C" "C" "T" "T"
#>  [3835] "A" "G" "A" "G" "G" "A" "G" "C" "T" "A" "G" "G" "G" "G" "A" "T" "G" "G"
#>  [3853] "A" "C" "T" "T" "G" "C" "A" "A" "T" "G" "G" "G" "C" "A" "T" "C" "A" "T"
#>  [3871] "G" "A" "T" "G" "T" "T" "G" "A" "A" "A" "T" "T" "A" "C" "T" "G" "A" "C"
#>  [3889] "T" "G" "A" "T" "T" "T" "T" "C" "A" "G" "T" "C" "A" "C" "A" "T" "C" "A"
#>  [3907] "G" "C" "T" "A" "T" "G" "G" "G" "C" "T" "A" "C" "C" "T" "T" "G" "C" "T"
#>  [3925] "G" "T" "C" "T" "T" "T" "A" "A" "C" "A" "T" "T" "T" "G" "T" "C" "A" "A"
#>  [3943] "A" "A" "C" "A" "A" "C" "T" "T" "T" "T" "T" "C" "A" "T" "T" "G" "C" "A"
#>  [3961] "C" "T" "A" "T" "G" "C" "A" "T" "G" "G" "A" "A" "G" "A" "C" "A" "A" "T"
#>  [3979] "G" "G" "C" "T" "A" "T" "G" "A" "T" "A" "C" "T" "G" "T" "C" "A" "A" "T"
#>  [3997] "T" "G" "T" "A" "T" "C" "T" "C" "T" "C" "T" "T" "C" "C" "C" "T" "T" "T"
#>  [4015] "A" "T" "G" "C" "C" "T" "G" "T" "C" "C" "A" "C" "G" "A" "C" "T" "T" "C"
#>  [4033] "T" "C" "A" "A" "A" "A" "A" "A" "C" "A" "A" "C" "A" "T" "G" "G" "C" "T"
#>  [4051] "T" "C" "C" "G" "G" "T" "G" "T" "T" "G" "C" "T" "G" "G" "G" "A" "T" "C"
#>  [4069] "T" "C" "T" "T" "G" "G" "A" "T" "G" "C" "A" "A" "A" "C" "C" "A" "C" "T"
#>  [4087] "A" "A" "C" "C" "A" "T" "G" "T" "T" "T" "C" "T" "T" "A" "T" "A" "A" "C"
#>  [4105] "A" "G" "A" "A" "A" "A" "C" "A" "A" "A" "A" "T" "C" "T" "G" "G" "G" "G"
#>  [4123] "A" "A" "G" "G" "A" "A" "A" "A" "G" "C" "T" "G" "G" "C" "C" "T" "C" "T"
#>  [4141] "C" "A" "A" "T" "G" "A" "A" "G" "G" "A" "A" "T" "T" "A" "T" "G" "G" "C"
#>  [4159] "T" "G" "T" "T" "G" "G" "A" "A" "T" "A" "G" "T" "T" "A" "G" "C" "A" "T"
#>  [4177] "T" "C" "T" "T" "C" "T" "A" "A" "G" "T" "T" "C" "A" "C" "T" "T" "C" "T"
#>  [4195] "C" "A" "A" "G" "A" "A" "T" "G" "A" "T" "G" "T" "G" "C" "C" "A" "C" "T"
#>  [4213] "A" "G" "C" "T" "G" "G" "C" "C" "C" "A" "C" "T" "A" "A" "T" "A" "G" "C"
#>  [4231] "T" "G" "G" "A" "G" "G" "C" "A" "T" "G" "C" "T" "A" "A" "T" "A" "G" "C"
#>  [4249] "A" "T" "G" "T" "T" "A" "T" "G" "T" "C" "A" "T" "A" "T" "C" "T" "G" "G"
#>  [4267] "A" "A" "G" "C" "T" "C" "G" "G" "C" "C" "G" "A" "T" "T" "T" "A" "T" "C"
#>  [4285] "A" "C" "T" "G" "G" "A" "G" "A" "A" "A" "G" "C" "G" "G" "C" "T" "G" "A"
#>  [4303] "G" "G" "T" "C" "T" "C" "C" "T" "G" "G" "G" "A" "A" "G" "A" "A" "G" "A"
#>  [4321] "A" "G" "C" "A" "G" "A" "A" "C" "A" "C" "T" "C" "T" "G" "G" "T" "G" "C"
#>  [4339] "C" "T" "C" "A" "C" "A" "C" "A" "A" "C" "A" "T" "A" "C" "T" "A" "G" "T"
#>  [4357] "G" "G" "A" "G" "G" "T" "C" "C" "A" "A" "G" "A" "T" "G" "A" "T" "G" "G"
#>  [4375] "A" "A" "C" "C" "A" "T" "G" "A" "A" "G" "A" "T" "A" "A" "A" "G" "G" "A"
#>  [4393] "T" "G" "A" "A" "G" "A" "G" "A" "G" "A" "G" "A" "T" "G" "A" "C" "A" "C"
#>  [4411] "A" "C" "T" "C" "A" "C" "C" "A" "T" "T" "C" "T" "C" "C" "T" "C" "A" "A"
#>  [4429] "A" "G" "C" "A" "A" "C" "T" "C" "T" "G" "C" "T" "A" "G" "C" "A" "A" "T"
#>  [4447] "C" "T" "C" "A" "G" "G" "G" "G" "T" "A" "T" "A" "C" "C" "C" "A" "A" "T"
#>  [4465] "G" "T" "C" "A" "A" "T" "A" "C" "C" "G" "G" "C" "G" "A" "C" "C" "C" "T"
#>  [4483] "C" "T" "T" "T" "G" "T" "G" "T" "G" "G" "T" "A" "T" "T" "T" "T" "T" "G"
#>  [4501] "G" "C" "A" "G" "A" "A" "A" "A" "A" "G" "A" "A" "A" "C" "A" "G" "A" "G"
#>  [4519] "A" "T" "C" "A" "G" "G" "A" "G" "T" "G" "C" "T" "A" "T" "G" "G" "G" "A"
#>  [4537] "C" "A" "C" "A" "C" "C" "C" "A" "G" "C" "C" "C" "T" "C" "C" "A" "G" "A"
#>  [4555] "A" "G" "T" "G" "G" "A" "A" "A" "G" "A" "G" "C" "A" "G" "T" "C" "C" "T"
#>  [4573] "T" "G" "A" "T" "G" "A" "T" "G" "G" "C" "A" "T" "T" "T" "A" "T" "A" "G"
#>  [4591] "A" "A" "T" "T" "C" "T" "C" "C" "A" "A" "A" "G" "A" "G" "G" "A" "T" "T"
#>  [4609] "G" "T" "T" "G" "G" "G" "C" "A" "G" "G" "T" "C" "T" "C" "A" "A" "G" "T"
#>  [4627] "A" "G" "G" "A" "G" "T" "A" "G" "G" "A" "G" "T" "T" "T" "T" "T" "C" "A"
#>  [4645] "A" "G" "A" "A" "G" "G" "C" "G" "T" "G" "T" "T" "C" "C" "A" "C" "A" "C"
#>  [4663] "A" "A" "T" "G" "T" "G" "G" "C" "A" "C" "G" "T" "C" "A" "C" "C" "A" "G"
#>  [4681] "G" "G" "G" "A" "G" "C" "T" "G" "T" "C" "C" "T" "C" "A" "T" "G" "T" "A"
#>  [4699] "C" "C" "A" "A" "G" "G" "G" "A" "A" "G" "A" "G" "A" "C" "T" "G" "G" "A"
#>  [4717] "A" "C" "C" "A" "A" "G" "T" "T" "G" "G" "G" "C" "C" "A" "G" "T" "G" "T"
#>  [4735] "C" "A" "A" "A" "A" "A" "A" "G" "A" "C" "T" "T" "G" "A" "T" "C" "T" "C"
#>  [4753] "A" "T" "A" "T" "G" "G" "A" "G" "G" "A" "G" "G" "T" "T" "G" "G" "A" "G"
#>  [4771] "G" "T" "T" "T" "C" "A" "A" "G" "G" "A" "T" "C" "C" "T" "G" "G" "A" "A"
#>  [4789] "C" "G" "C" "G" "G" "G" "A" "G" "A" "A" "G" "A" "A" "G" "T" "G" "C" "A"
#>  [4807] "G" "G" "T" "G" "A" "T" "T" "G" "C" "T" "G" "T" "T" "G" "A" "A" "C" "C"
#>  [4825] "G" "G" "G" "G" "A" "A" "G" "A" "A" "C" "C" "C" "C" "A" "A" "A" "A" "A"
#>  [4843] "T" "G" "T" "A" "C" "A" "G" "A" "C" "A" "G" "C" "G" "C" "C" "G" "G" "G"
#>  [4861] "T" "A" "C" "C" "T" "T" "C" "A" "A" "G" "A" "C" "C" "C" "C" "T" "G" "A"
#>  [4879] "A" "G" "G" "C" "G" "A" "A" "G" "T" "T" "G" "G" "A" "G" "C" "C" "A" "T"
#>  [4897] "A" "G" "C" "T" "C" "T" "A" "G" "A" "C" "T" "T" "T" "A" "A" "A" "C" "C"
#>  [4915] "C" "G" "G" "C" "A" "C" "A" "T" "C" "T" "G" "G" "A" "T" "C" "T" "C" "C"
#>  [4933] "T" "A" "T" "C" "G" "T" "G" "A" "A" "C" "A" "G" "A" "G" "A" "G" "G" "G"
#>  [4951] "A" "A" "A" "A" "A" "T" "A" "G" "T" "A" "G" "G" "T" "C" "T" "T" "T" "A"
#>  [4969] "T" "G" "G" "A" "A" "A" "T" "G" "G" "A" "G" "T" "G" "G" "T" "G" "A" "C"
#>  [4987] "A" "A" "C" "A" "A" "G" "T" "G" "G" "T" "A" "C" "C" "T" "A" "C" "G" "T"
#>  [5005] "C" "A" "G" "T" "G" "C" "C" "A" "T" "A" "G" "C" "T" "C" "A" "A" "G" "C"
#>  [5023] "T" "A" "A" "A" "G" "C" "A" "T" "C" "A" "C" "A" "A" "G" "A" "A" "G" "G"
#>  [5041] "G" "C" "C" "T" "C" "T" "A" "C" "C" "A" "G" "A" "G" "A" "T" "T" "G" "A"
#>  [5059] "G" "G" "A" "C" "G" "A" "G" "G" "T" "G" "T" "T" "T" "A" "G" "G" "A" "A"
#>  [5077] "A" "A" "G" "A" "A" "A" "C" "T" "T" "A" "A" "C" "A" "A" "T" "A" "A" "T"
#>  [5095] "G" "G" "A" "C" "C" "T" "A" "C" "A" "T" "C" "C" "A" "G" "G" "A" "T" "C"
#>  [5113] "G" "G" "G" "A" "A" "A" "A" "A" "C" "A" "A" "G" "A" "A" "G" "A" "T" "A"
#>  [5131] "C" "C" "T" "T" "C" "C" "A" "G" "C" "C" "A" "T" "A" "G" "T" "C" "C" "G"
#>  [5149] "T" "G" "A" "G" "G" "C" "C" "A" "T" "A" "A" "A" "A" "A" "G" "A" "A" "A"
#>  [5167] "G" "C" "T" "G" "C" "G" "C" "A" "C" "G" "C" "T" "A" "G" "T" "C" "T" "T"
#>  [5185] "A" "G" "C" "T" "C" "C" "C" "A" "C" "A" "A" "G" "A" "G" "T" "T" "G" "T"
#>  [5203] "C" "G" "C" "T" "T" "C" "T" "G" "A" "A" "A" "T" "G" "G" "C" "A" "G" "A"
#>  [5221] "G" "G" "C" "G" "C" "T" "C" "A" "A" "G" "G" "G" "A" "A" "T" "G" "C" "C"
#>  [5239] "A" "A" "T" "A" "A" "G" "G" "T" "A" "T" "C" "A" "G" "A" "C" "A" "A" "C"
#>  [5257] "A" "G" "C" "A" "G" "T" "G" "A" "A" "G" "A" "G" "T" "G" "A" "A" "C" "A"
#>  [5275] "C" "A" "C" "G" "G" "G" "A" "A" "A" "G" "G" "A" "G" "A" "T" "A" "G" "T"
#>  [5293] "T" "G" "A" "C" "C" "T" "T" "A" "T" "G" "T" "G" "T" "C" "A" "C" "G" "C"
#>  [5311] "C" "A" "C" "T" "T" "T" "C" "A" "C" "T" "A" "T" "G" "C" "G" "T" "C" "T"
#>  [5329] "C" "C" "T" "G" "T" "C" "T" "C" "C" "T" "G" "T" "G" "A" "G" "A" "G" "T"
#>  [5347] "T" "C" "C" "C" "A" "A" "T" "T" "A" "T" "A" "A" "T" "A" "T" "G" "A" "T"
#>  [5365] "T" "A" "T" "C" "A" "T" "G" "G" "A" "T" "G" "A" "A" "G" "C" "A" "C" "A"
#>  [5383] "T" "T" "T" "T" "A" "C" "C" "G" "A" "T" "C" "C" "A" "G" "C" "C" "A" "G"
#>  [5401] "C" "A" "T" "A" "G" "C" "A" "G" "C" "C" "A" "G" "A" "G" "G" "G" "T" "A"
#>  [5419] "T" "A" "T" "C" "T" "C" "A" "A" "C" "C" "C" "G" "A" "G" "T" "G" "G" "G"
#>  [5437] "T" "A" "T" "G" "G" "G" "T" "G" "A" "A" "G" "C" "A" "G" "C" "T" "G" "C"
#>  [5455] "G" "A" "T" "T" "T" "T" "C" "A" "T" "G" "A" "C" "A" "G" "C" "C" "A" "C"
#>  [5473] "T" "C" "C" "C" "C" "C" "C" "G" "G" "A" "T" "C" "G" "G" "T" "G" "G" "A"
#>  [5491] "G" "G" "C" "C" "T" "T" "T" "C" "C" "A" "C" "A" "G" "A" "G" "C" "A" "A"
#>  [5509] "T" "G" "C" "A" "G" "T" "T" "A" "T" "C" "C" "A" "A" "G" "A" "T" "G" "A"
#>  [5527] "G" "G" "A" "A" "A" "G" "A" "G" "A" "C" "A" "T" "T" "C" "C" "T" "G" "A"
#>  [5545] "A" "A" "G" "A" "T" "C" "A" "T" "G" "G" "A" "A" "C" "T" "C" "A" "G" "G"
#>  [5563] "C" "T" "A" "T" "G" "A" "C" "T" "G" "G" "A" "T" "C" "A" "C" "T" "G" "A"
#>  [5581] "T" "T" "T" "C" "C" "C" "A" "G" "G" "T" "A" "A" "A" "A" "C" "A" "G" "T"
#>  [5599] "C" "T" "G" "G" "T" "T" "T" "G" "T" "T" "C" "C" "A" "A" "G" "C" "A" "T"
#>  [5617] "C" "A" "A" "A" "T" "C" "A" "G" "G" "A" "A" "A" "T" "G" "A" "C" "A" "T"
#>  [5635] "T" "G" "C" "C" "A" "A" "C" "T" "G" "T" "T" "T" "A" "A" "G" "A" "A" "A"
#>  [5653] "G" "A" "A" "T" "G" "G" "G" "A" "A" "A" "C" "G" "G" "G" "T" "G" "G" "T"
#>  [5671] "C" "C" "A" "A" "T" "T" "G" "A" "G" "C" "A" "G" "A" "A" "A" "A" "A" "C"
#>  [5689] "T" "T" "T" "T" "G" "A" "C" "A" "C" "T" "G" "A" "G" "T" "A" "C" "C" "A"
#>  [5707] "G" "A" "A" "A" "A" "C" "A" "A" "A" "A" "A" "A" "T" "A" "A" "C" "G" "A"
#>  [5725] "C" "T" "G" "G" "G" "A" "C" "T" "A" "T" "G" "T" "T" "G" "T" "C" "A" "C"
#>  [5743] "A" "A" "C" "A" "G" "A" "C" "A" "T" "A" "T" "C" "C" "G" "A" "A" "A" "T"
#>  [5761] "G" "G" "G" "A" "G" "C" "A" "A" "A" "C" "T" "T" "C" "C" "G" "A" "G" "C"
#>  [5779] "C" "G" "A" "C" "A" "G" "G" "G" "T" "A" "A" "T" "A" "G" "A" "C" "C" "C"
#>  [5797] "G" "A" "G" "G" "C" "G" "G" "T" "G" "C" "C" "T" "G" "A" "A" "A" "C" "C"
#>  [5815] "G" "G" "T" "A" "A" "T" "A" "C" "T" "A" "A" "A" "A" "G" "A" "T" "G" "G"
#>  [5833] "C" "C" "C" "A" "G" "A" "G" "C" "G" "T" "G" "T" "C" "A" "T" "T" "C" "T"
#>  [5851] "A" "G" "C" "C" "G" "G" "A" "C" "C" "G" "A" "T" "G" "C" "C" "A" "G" "T"
#>  [5869] "G" "A" "C" "T" "G" "T" "G" "G" "C" "T" "A" "G" "C" "G" "C" "C" "G" "C"
#>  [5887] "C" "C" "A" "G" "A" "G" "G" "A" "G" "A" "G" "G" "A" "A" "G" "A" "A" "T"
#>  [5905] "T" "G" "G" "A" "A" "G" "G" "A" "A" "C" "C" "A" "A" "A" "A" "T" "A" "A"
#>  [5923] "G" "G" "A" "A" "G" "G" "C" "G" "A" "T" "C" "A" "G" "T" "A" "T" "A" "T"
#>  [5941] "T" "T" "A" "C" "A" "T" "G" "G" "G" "A" "C" "A" "G" "C" "C" "T" "C" "T"
#>  [5959] "A" "A" "A" "C" "A" "A" "T" "G" "A" "T" "G" "A" "G" "G" "A" "C" "C" "A"
#>  [5977] "C" "G" "C" "C" "C" "A" "T" "T" "G" "G" "A" "C" "A" "G" "A" "A" "G" "C"
#>  [5995] "A" "A" "A" "A" "A" "T" "G" "C" "T" "C" "C" "T" "T" "G" "A" "C" "A" "A"
#>  [6013] "C" "A" "T" "A" "A" "A" "C" "A" "C" "A" "C" "C" "A" "G" "A" "A" "G" "G"
#>  [6031] "G" "A" "T" "T" "A" "T" "C" "C" "C" "A" "G" "C" "C" "C" "T" "C" "T" "T"
#>  [6049] "T" "G" "A" "G" "C" "C" "G" "G" "A" "G" "A" "G" "A" "G" "A" "A" "A" "A"
#>  [6067] "G" "A" "G" "T" "G" "C" "A" "G" "C" "A" "A" "T" "A" "G" "A" "C" "G" "G"
#>  [6085] "G" "G" "A" "A" "T" "A" "C" "A" "G" "A" "C" "T" "A" "C" "G" "G" "G" "G"
#>  [6103] "T" "G" "A" "A" "G" "C" "G" "A" "G" "G" "A" "A" "A" "A" "C" "G" "T" "T"
#>  [6121] "C" "G" "T" "G" "G" "A" "G" "C" "T" "C" "A" "T" "G" "A" "G" "A" "A" "G"
#>  [6139] "A" "G" "G" "A" "G" "A" "T" "C" "T" "A" "C" "C" "T" "G" "T" "C" "T" "G"
#>  [6157] "G" "C" "T" "A" "T" "C" "C" "T" "A" "C" "A" "A" "A" "G" "T" "T" "G" "C"
#>  [6175] "C" "T" "C" "A" "G" "A" "A" "G" "G" "C" "T" "T" "C" "C" "A" "G" "T" "A"
#>  [6193] "C" "T" "C" "C" "G" "A" "C" "A" "G" "A" "A" "G" "G" "T" "G" "G" "T" "G"
#>  [6211] "C" "T" "T" "T" "G" "A" "T" "G" "G" "G" "G" "A" "A" "A" "G" "G" "A" "A"
#>  [6229] "C" "A" "A" "C" "C" "A" "G" "G" "T" "G" "T" "T" "G" "G" "A" "G" "G" "A"
#>  [6247] "G" "A" "A" "C" "A" "T" "G" "G" "A" "C" "G" "T" "G" "G" "A" "G" "A" "T"
#>  [6265] "C" "T" "G" "G" "A" "C" "A" "A" "A" "A" "G" "A" "A" "G" "G" "A" "G" "A"
#>  [6283] "A" "A" "G" "A" "A" "A" "G" "A" "A" "A" "C" "T" "A" "C" "G" "A" "C" "C"
#>  [6301] "C" "C" "G" "C" "T" "G" "G" "C" "T" "G" "G" "A" "T" "G" "C" "C" "A" "G"
#>  [6319] "A" "A" "C" "A" "T" "A" "C" "T" "C" "T" "G" "A" "C" "C" "C" "A" "C" "T"
#>  [6337] "G" "G" "C" "T" "C" "T" "G" "C" "G" "C" "G" "A" "A" "T" "T" "C" "A" "A"
#>  [6355] "A" "G" "A" "G" "T" "T" "C" "G" "C" "A" "G" "C" "A" "G" "G" "A" "A" "G"
#>  [6373] "A" "A" "G" "A" "A" "G" "C" "G" "T" "C" "T" "C" "A" "G" "G" "T" "G" "A"
#>  [6391] "C" "C" "T" "A" "A" "T" "A" "T" "T" "A" "G" "A" "A" "A" "T" "A" "G" "G"
#>  [6409] "G" "A" "A" "A" "C" "T" "T" "C" "C" "A" "C" "A" "A" "C" "A" "T" "T" "T"
#>  [6427] "A" "A" "C" "G" "C" "A" "A" "A" "G" "G" "G" "C" "C" "C" "A" "G" "A" "A"
#>  [6445] "C" "G" "C" "C" "T" "T" "G" "G" "A" "C" "A" "A" "T" "C" "T" "G" "G" "T"
#>  [6463] "T" "A" "T" "G" "T" "T" "G" "C" "A" "C" "A" "A" "C" "T" "C" "T" "G" "A"
#>  [6481] "A" "C" "A" "A" "G" "G" "A" "G" "G" "A" "A" "A" "A" "G" "C" "C" "T" "A"
#>  [6499] "T" "A" "G" "A" "C" "A" "C" "G" "C" "C" "A" "T" "G" "G" "A" "A" "G" "A"
#>  [6517] "A" "C" "T" "A" "C" "C" "A" "G" "A" "C" "A" "C" "C" "A" "T" "A" "G" "A"
#>  [6535] "A" "A" "C" "G" "T" "T" "A" "A" "T" "G" "C" "T" "C" "C" "T" "A" "G" "C"
#>  [6553] "T" "T" "T" "G" "A" "T" "A" "G" "C" "T" "G" "T" "G" "C" "T" "G" "A" "C"
#>  [6571] "T" "G" "G" "T" "G" "G" "A" "G" "T" "G" "A" "C" "G" "T" "T" "G" "T" "T"
#>  [6589] "C" "T" "T" "C" "C" "T" "A" "T" "C" "A" "G" "G" "A" "A" "G" "G" "G" "G"
#>  [6607] "T" "C" "T" "A" "G" "G" "A" "A" "A" "A" "A" "C" "A" "T" "C" "C" "A" "T"
#>  [6625] "T" "G" "G" "C" "C" "T" "A" "C" "T" "C" "T" "G" "C" "G" "T" "G" "A" "T"
#>  [6643] "T" "G" "C" "C" "T" "C" "A" "A" "G" "T" "G" "C" "A" "C" "T" "G" "T" "T"
#>  [6661] "A" "T" "G" "G" "A" "T" "G" "G" "C" "C" "A" "G" "T" "G" "T" "G" "G" "A"
#>  [6679] "A" "C" "C" "C" "C" "A" "T" "T" "G" "G" "A" "T" "A" "G" "C" "G" "G" "C"
#>  [6697] "C" "T" "C" "T" "A" "T" "C" "A" "T" "A" "C" "T" "G" "G" "A" "G" "T" "T"
#>  [6715] "C" "T" "T" "T" "C" "T" "G" "A" "T" "G" "G" "T" "G" "T" "T" "G" "C" "T"
#>  [6733] "T" "A" "T" "T" "C" "C" "A" "G" "A" "G" "C" "C" "G" "G" "A" "C" "A" "G"
#>  [6751] "A" "C" "A" "G" "C" "G" "C" "A" "C" "T" "C" "C" "A" "C" "A" "A" "G" "A"
#>  [6769] "C" "A" "A" "C" "C" "A" "G" "C" "T" "A" "G" "C" "A" "T" "A" "C" "G" "T"
#>  [6787] "G" "G" "T" "G" "A" "T" "A" "G" "G" "T" "C" "T" "G" "T" "T" "A" "T" "T"
#>  [6805] "C" "A" "T" "G" "A" "T" "A" "T" "T" "G" "A" "C" "A" "G" "T" "G" "G" "C"
#>  [6823] "A" "G" "C" "C" "A" "A" "T" "G" "A" "G" "A" "T" "G" "G" "G" "A" "T" "T"
#>  [6841] "A" "C" "T" "G" "G" "A" "A" "A" "C" "C" "A" "C" "A" "A" "A" "G" "A" "A"
#>  [6859] "G" "G" "A" "C" "C" "T" "G" "G" "G" "G" "A" "T" "T" "G" "G" "T" "C" "A"
#>  [6877] "T" "G" "C" "A" "G" "C" "T" "G" "C" "T" "G" "A" "A" "A" "A" "C" "C" "A"
#>  [6895] "C" "C" "A" "T" "C" "A" "T" "G" "C" "T" "G" "C" "A" "A" "T" "G" "C" "T"
#>  [6913] "G" "G" "A" "C" "G" "T" "A" "G" "A" "C" "C" "T" "A" "C" "A" "T" "C" "C"
#>  [6931] "A" "G" "C" "T" "T" "C" "A" "G" "C" "C" "T" "G" "G" "A" "C" "T" "C" "T"
#>  [6949] "C" "T" "A" "T" "G" "C" "A" "G" "T" "G" "G" "C" "C" "A" "C" "A" "A" "C"
#>  [6967] "A" "A" "T" "T" "A" "T" "C" "A" "C" "T" "C" "C" "C" "A" "T" "G" "A" "T"
#>  [6985] "G" "A" "G" "A" "C" "A" "C" "A" "C" "A" "A" "T" "T" "G" "A" "A" "A" "A"
#>  [7003] "C" "A" "C" "A" "A" "C" "G" "G" "C" "A" "A" "A" "T" "A" "T" "T" "T" "C"
#>  [7021] "C" "C" "T" "G" "A" "C" "A" "G" "C" "T" "A" "T" "T" "G" "C" "A" "A" "A"
#>  [7039] "C" "C" "A" "G" "G" "C" "A" "G" "C" "T" "A" "T" "A" "T" "T" "G" "A" "T"
#>  [7057] "G" "G" "G" "A" "C" "T" "T" "G" "A" "C" "A" "A" "G" "G" "G" "A" "T" "G"
#>  [7075] "G" "C" "C" "A" "A" "T" "A" "T" "C" "A" "A" "A" "G" "A" "T" "G" "G" "A"
#>  [7093] "C" "A" "T" "A" "G" "G" "A" "G" "T" "T" "C" "C" "A" "C" "T" "T" "C" "T"
#>  [7111] "C" "G" "C" "C" "T" "T" "G" "G" "G" "G" "T" "G" "C" "T" "A" "T" "T" "C"
#>  [7129] "T" "C" "A" "G" "G" "T" "G" "A" "A" "C" "C" "C" "G" "C" "T" "G" "A" "C"
#>  [7147] "G" "C" "T" "G" "A" "C" "A" "G" "C" "G" "G" "C" "G" "G" "T" "A" "T" "T"
#>  [7165] "G" "A" "T" "G" "C" "T" "A" "G" "T" "G" "G" "C" "T" "C" "A" "T" "T" "A"
#>  [7183] "T" "G" "C" "C" "A" "T" "A" "A" "T" "T" "G" "G" "A" "C" "C" "C" "G" "G"
#>  [7201] "A" "C" "T" "G" "C" "A" "A" "G" "C" "A" "A" "A" "A" "G" "C" "T" "A" "C"
#>  [7219] "T" "A" "G" "A" "G" "A" "A" "G" "C" "T" "C" "A" "A" "A" "A" "A" "A" "G"
#>  [7237] "G" "A" "C" "A" "G" "C" "A" "G" "C" "C" "G" "G" "A" "A" "T" "A" "A" "T"
#>  [7255] "G" "A" "A" "A" "A" "A" "C" "C" "C" "A" "A" "C" "T" "G" "T" "C" "G" "A"
#>  [7273] "C" "G" "G" "G" "A" "T" "C" "G" "T" "T" "G" "C" "A" "A" "T" "A" "G" "A"
#>  [7291] "T" "T" "T" "G" "G" "A" "C" "C" "C" "T" "G" "T" "G" "G" "T" "T" "T" "A"
#>  [7309] "C" "G" "A" "T" "G" "C" "A" "A" "A" "A" "T" "T" "T" "G" "A" "A" "A" "A"
#>  [7327] "A" "C" "A" "G" "C" "T" "A" "G" "G" "C" "C" "A" "A" "A" "T" "A" "A" "T"
#>  [7345] "G" "T" "T" "G" "T" "T" "G" "A" "T" "A" "C" "T" "T" "T" "G" "C" "A" "C"
#>  [7363] "A" "T" "C" "A" "C" "A" "G" "A" "T" "C" "C" "T" "C" "C" "T" "G" "A" "T"
#>  [7381] "G" "C" "G" "G" "A" "C" "C" "A" "C" "A" "T" "G" "G" "G" "C" "C" "T" "T"
#>  [7399] "G" "T" "G" "T" "G" "A" "A" "T" "C" "C" "A" "T" "C" "A" "C" "A" "C" "T"
#>  [7417] "A" "G" "C" "C" "A" "C" "T" "G" "G" "A" "C" "C" "T" "C" "T" "G" "A" "C"
#>  [7435] "T" "A" "C" "G" "C" "T" "T" "T" "G" "G" "G" "A" "G" "G" "G" "A" "T" "C"
#>  [7453] "T" "C" "C" "A" "G" "G" "A" "A" "A" "A" "T" "T" "C" "T" "G" "G" "A" "A"
#>  [7471] "C" "A" "C" "C" "A" "C" "G" "A" "T" "A" "G" "C" "G" "G" "T" "G" "T" "C"
#>  [7489] "C" "A" "T" "G" "G" "C" "A" "A" "A" "C" "A" "T" "T" "T" "T" "T" "A" "G"
#>  [7507] "G" "G" "G" "A" "A" "G" "T" "T" "A" "T" "C" "T" "A" "G" "C" "A" "G" "G"
#>  [7525] "A" "G" "C" "A" "G" "G" "T" "C" "T" "G" "G" "C" "C" "T" "T" "T" "T" "C"
#>  [7543] "A" "T" "T" "A" "A" "T" "G" "A" "A" "A" "T" "C" "T" "C" "T" "A" "G" "G"
#>  [7561] "A" "G" "G" "A" "G" "G" "T" "A" "G" "G" "A" "G" "A" "G" "G" "C" "A" "C"
#>  [7579] "G" "G" "G" "A" "G" "C" "C" "C" "A" "A" "G" "G" "G" "G" "A" "A" "A" "C"
#>  [7597] "A" "C" "T" "G" "G" "G" "A" "G" "A" "A" "A" "A" "A" "T" "G" "G" "A" "A"
#>  [7615] "A" "A" "G" "A" "C" "A" "G" "C" "T" "A" "A" "A" "C" "C" "A" "A" "T" "T"
#>  [7633] "G" "A" "G" "C" "A" "A" "G" "T" "C" "A" "G" "A" "A" "T" "T" "C" "A" "A"
#>  [7651] "C" "A" "C" "T" "T" "A" "C" "A" "A" "A" "A" "G" "G" "A" "G" "T" "G" "G"
#>  [7669] "G" "A" "T" "T" "A" "T" "A" "G" "A" "G" "G" "T" "G" "G" "A" "T" "A" "G"
#>  [7687] "A" "T" "C" "T" "G" "A" "A" "G" "C" "C" "A" "A" "A" "G" "A" "G" "G" "G"
#>  [7705] "G" "T" "T" "A" "A" "A" "A" "A" "G" "A" "G" "G" "A" "G" "A" "A" "A" "C"
#>  [7723] "G" "A" "C" "T" "A" "A" "A" "C" "A" "C" "G" "C" "A" "G" "T" "G" "T" "C"
#>  [7741] "G" "A" "G" "A" "G" "G" "A" "A" "C" "G" "G" "C" "C" "A" "A" "A" "C" "T"
#>  [7759] "G" "A" "G" "G" "T" "G" "G" "T" "T" "T" "G" "T" "G" "G" "A" "G" "A" "G"
#>  [7777] "G" "A" "A" "C" "C" "T" "T" "G" "T" "G" "A" "A" "A" "C" "C" "A" "G" "A"
#>  [7795] "A" "G" "G" "G" "A" "A" "A" "G" "T" "C" "A" "T" "A" "G" "A" "C" "C" "T"
#>  [7813] "C" "G" "G" "T" "T" "G" "T" "G" "G" "A" "A" "G" "A" "G" "G" "T" "G" "G"
#>  [7831] "C" "T" "G" "G" "T" "C" "A" "T" "A" "T" "T" "A" "T" "T" "G" "C" "G" "C"
#>  [7849] "T" "G" "G" "G" "C" "T" "G" "A" "A" "G" "A" "A" "A" "G" "T" "C" "A" "C"
#>  [7867] "A" "G" "A" "A" "G" "T" "G" "A" "A" "A" "G" "G" "A" "T" "A" "C" "A" "C"
#>  [7885] "G" "A" "A" "A" "G" "G" "A" "G" "G" "A" "C" "C" "T" "G" "G" "A" "C" "A"
#>  [7903] "T" "G" "A" "G" "G" "A" "A" "C" "C" "A" "A" "T" "C" "C" "C" "A" "A" "T"
#>  [7921] "G" "G" "C" "A" "A" "C" "C" "T" "A" "T" "G" "G" "A" "T" "G" "G" "A" "A"
#>  [7939] "C" "C" "T" "A" "G" "T" "A" "A" "A" "G" "C" "T" "A" "T" "A" "C" "T" "C"
#>  [7957] "C" "G" "G" "G" "A" "A" "A" "G" "A" "T" "G" "T" "A" "T" "T" "C" "T" "T"
#>  [7975] "T" "A" "C" "A" "C" "C" "A" "C" "C" "T" "G" "A" "G" "A" "A" "A" "T" "G"
#>  [7993] "T" "G" "A" "C" "A" "C" "C" "C" "T" "C" "T" "T" "G" "T" "G" "T" "G" "A"
#>  [8011] "T" "A" "T" "T" "G" "G" "T" "G" "A" "G" "T" "C" "C" "T" "C" "T" "C" "C"
#>  [8029] "G" "A" "A" "C" "C" "C" "A" "A" "C" "T" "A" "T" "A" "G" "A" "A" "G" "A"
#>  [8047] "A" "G" "G" "A" "A" "G" "A" "A" "C" "G" "T" "T" "A" "C" "G" "T" "G" "T"
#>  [8065] "T" "C" "T" "A" "A" "A" "G" "A" "T" "G" "G" "T" "G" "G" "A" "A" "C" "C"
#>  [8083] "A" "T" "G" "G" "C" "T" "C" "A" "G" "A" "G" "G" "A" "A" "A" "C" "C" "A"
#>  [8101] "A" "T" "T" "T" "T" "G" "C" "A" "T" "A" "A" "A" "A" "A" "T" "T" "C" "T"
#>  [8119] "A" "A" "A" "T" "C" "C" "C" "T" "A" "T" "A" "T" "G" "C" "C" "G" "A" "G"
#>  [8137] "T" "G" "T" "G" "G" "T" "A" "G" "A" "A" "A" "C" "T" "T" "T" "G" "G" "A"
#>  [8155] "G" "C" "A" "A" "A" "T" "G" "C" "A" "A" "A" "G" "A" "A" "A" "A" "C" "A"
#>  [8173] "T" "G" "G" "A" "G" "G" "A" "A" "T" "G" "C" "T" "A" "G" "T" "G" "C" "G"
#>  [8191] "A" "A" "A" "T" "C" "C" "A" "C" "T" "C" "T" "C" "A" "A" "G" "A" "A" "A"
#>  [8209] "C" "T" "C" "C" "A" "C" "T" "C" "A" "T" "G" "A" "A" "A" "T" "G" "T" "A"
#>  [8227] "C" "T" "G" "G" "G" "T" "T" "T" "C" "A" "T" "G" "T" "G" "G" "A" "A" "C"
#>  [8245] "A" "G" "G" "A" "A" "A" "C" "A" "T" "T" "G" "T" "G" "T" "C" "A" "G" "C"
#>  [8263] "A" "G" "T" "A" "A" "A" "C" "A" "T" "G" "A" "C" "A" "T" "C" "T" "A" "G"
#>  [8281] "A" "A" "T" "G" "C" "T" "G" "C" "T" "A" "A" "A" "T" "C" "G" "A" "T" "T"
#>  [8299] "C" "A" "C" "A" "A" "T" "G" "G" "C" "T" "C" "A" "C" "A" "G" "G" "A" "A"
#>  [8317] "G" "C" "C" "A" "A" "C" "A" "T" "A" "T" "G" "A" "A" "A" "G" "A" "G" "A"
#>  [8335] "C" "G" "T" "G" "G" "A" "C" "T" "T" "A" "G" "G" "C" "G" "C" "T" "G" "G"
#>  [8353] "A" "A" "C" "A" "A" "G" "A" "C" "A" "T" "G" "T" "G" "G" "C" "A" "G" "T"
#>  [8371] "A" "G" "A" "A" "C" "C" "A" "G" "A" "G" "G" "T" "G" "G" "C" "C" "A" "A"
#>  [8389] "C" "C" "T" "A" "G" "A" "T" "A" "T" "C" "A" "T" "T" "G" "G" "C" "C" "A"
#>  [8407] "G" "A" "G" "G" "A" "T" "A" "G" "A" "G" "A" "A" "T" "A" "T" "A" "A" "A"
#>  [8425] "A" "A" "A" "T" "G" "A" "A" "C" "A" "C" "A" "A" "A" "T" "C" "A" "A" "C"
#>  [8443] "A" "T" "G" "G" "C" "A" "T" "T" "A" "T" "G" "A" "T" "G" "A" "G" "G" "A"
#>  [8461] "C" "A" "A" "T" "C" "C" "A" "T" "A" "C" "A" "A" "A" "A" "C" "A" "T" "G"
#>  [8479] "G" "G" "C" "C" "T" "A" "T" "C" "A" "T" "G" "G" "A" "T" "C" "A" "T" "A"
#>  [8497] "T" "G" "A" "G" "G" "T" "C" "A" "A" "G" "C" "C" "A" "T" "C" "A" "G" "G"
#>  [8515] "A" "T" "C" "A" "G" "C" "C" "T" "C" "A" "T" "C" "C" "A" "T" "G" "G" "T"
#>  [8533] "C" "A" "A" "T" "G" "G" "T" "G" "T" "G" "G" "T" "G" "A" "G" "A" "C" "T"
#>  [8551] "G" "C" "T" "A" "A" "C" "C" "A" "A" "A" "C" "C" "A" "T" "G" "G" "G" "A"
#>  [8569] "T" "G" "T" "C" "A" "T" "T" "C" "C" "C" "A" "T" "G" "G" "T" "C" "A" "C"
#>  [8587] "A" "C" "A" "A" "A" "T" "A" "G" "C" "C" "A" "T" "G" "A" "C" "T" "G" "A"
#>  [8605] "C" "A" "C" "C" "A" "C" "A" "C" "C" "C" "T" "T" "T" "G" "G" "A" "C" "A"
#>  [8623] "A" "C" "A" "G" "A" "G" "G" "G" "T" "G" "T" "T" "T" "A" "A" "A" "G" "A"
#>  [8641] "G" "A" "A" "A" "G" "T" "T" "G" "A" "C" "A" "C" "G" "C" "G" "T" "A" "C"
#>  [8659] "A" "C" "C" "A" "A" "A" "A" "G" "C" "G" "A" "A" "A" "C" "G" "A" "G" "G"
#>  [8677] "C" "A" "C" "A" "G" "C" "A" "C" "A" "A" "A" "T" "T" "A" "T" "G" "G" "A"
#>  [8695] "G" "G" "T" "G" "A" "C" "A" "G" "C" "C" "A" "G" "G" "T" "G" "G" "T" "T"
#>  [8713] "A" "T" "G" "G" "G" "G" "T" "T" "T" "T" "C" "T" "C" "T" "C" "T" "A" "G"
#>  [8731] "A" "A" "A" "C" "A" "A" "A" "A" "A" "A" "C" "C" "C" "A" "G" "A" "A" "T"
#>  [8749] "C" "T" "G" "C" "A" "C" "A" "A" "G" "A" "G" "A" "G" "G" "A" "G" "T" "T"
#>  [8767] "C" "A" "C" "A" "A" "G" "A" "A" "A" "A" "G" "T" "C" "A" "G" "G" "T" "C"
#>  [8785] "A" "A" "A" "C" "G" "C" "A" "G" "C" "T" "A" "T" "T" "G" "G" "A" "G" "C"
#>  [8803] "A" "G" "T" "G" "T" "T" "C" "G" "T" "T" "G" "A" "T" "G" "A" "A" "A" "A"
#>  [8821] "T" "C" "A" "A" "T" "G" "G" "A" "A" "C" "T" "C" "A" "G" "C" "A" "A" "A"
#>  [8839] "A" "G" "A" "G" "G" "C" "A" "G" "T" "G" "G" "A" "A" "G" "A" "T" "G" "A"
#>  [8857] "A" "C" "G" "G" "T" "T" "C" "T" "G" "G" "G" "A" "C" "C" "T" "T" "G" "T"
#>  [8875] "G" "C" "A" "C" "A" "G" "A" "G" "A" "G" "A" "G" "G" "G" "A" "G" "C" "T"
#>  [8893] "T" "C" "A" "T" "A" "A" "A" "C" "A" "A" "G" "G" "A" "A" "A" "A" "T" "G"
#>  [8911] "T" "G" "C" "C" "A" "C" "G" "T" "G" "T" "G" "T" "C" "T" "A" "C" "A" "A"
#>  [8929] "C" "A" "T" "G" "A" "T" "G" "G" "G" "A" "A" "A" "G" "A" "G" "A" "G" "A"
#>  [8947] "G" "A" "A" "A" "A" "A" "A" "T" "T" "A" "G" "G" "A" "G" "A" "G" "T" "T"
#>  [8965] "C" "G" "G" "A" "A" "A" "G" "G" "C" "A" "A" "A" "A" "G" "G" "A" "A" "G"
#>  [8983] "T" "C" "G" "C" "G" "C" "A" "A" "T" "A" "T" "G" "G" "T" "A" "C" "A" "T"
#>  [9001] "G" "T" "G" "G" "T" "T" "G" "G" "G" "A" "G" "C" "G" "C" "G" "C" "T" "T"
#>  [9019] "T" "T" "T" "A" "G" "A" "G" "T" "T" "T" "G" "A" "A" "G" "C" "C" "C" "T"
#>  [9037] "T" "G" "G" "T" "T" "T" "C" "A" "T" "G" "A" "A" "T" "G" "A" "A" "G" "A"
#>  [9055] "T" "C" "A" "C" "T" "G" "G" "T" "T" "C" "A" "G" "C" "A" "G" "A" "G" "A"
#>  [9073] "G" "A" "A" "T" "T" "C" "A" "C" "T" "C" "A" "G" "T" "G" "G" "A" "G" "T"
#>  [9091] "G" "G" "A" "A" "G" "G" "A" "G" "A" "A" "G" "G" "A" "C" "T" "C" "C" "A"
#>  [9109] "C" "A" "A" "A" "C" "T" "T" "G" "G" "A" "T" "A" "C" "A" "T" "A" "C" "T"
#>  [9127] "C" "A" "G" "A" "G" "A" "C" "A" "T" "A" "T" "C" "A" "A" "A" "G" "A" "T"
#>  [9145] "T" "C" "C" "A" "G" "G" "G" "G" "G" "A" "A" "A" "T" "A" "T" "G" "T" "A"
#>  [9163] "T" "G" "C" "A" "G" "A" "T" "G" "A" "C" "A" "C" "A" "G" "C" "C" "G" "G"
#>  [9181] "A" "T" "G" "G" "G" "A" "C" "A" "C" "A" "A" "G" "A" "A" "T" "A" "A" "C"
#>  [9199] "A" "G" "A" "G" "G" "A" "T" "G" "A" "T" "C" "T" "T" "C" "A" "G" "A" "A"
#>  [9217] "T" "G" "A" "G" "G" "C" "C" "A" "A" "A" "A" "T" "C" "A" "C" "T" "G" "A"
#>  [9235] "C" "A" "T" "C" "A" "T" "G" "G" "A" "A" "C" "C" "T" "G" "A" "A" "C" "A"
#>  [9253] "T" "G" "C" "C" "C" "T" "A" "T" "T" "G" "G" "C" "C" "A" "C" "G" "T" "C"
#>  [9271] "A" "A" "T" "C" "T" "T" "T" "A" "A" "G" "C" "T" "A" "A" "C" "C" "T" "A"
#>  [9289] "C" "C" "A" "A" "A" "A" "C" "A" "A" "G" "G" "T" "A" "G" "T" "A" "A" "G"
#>  [9307] "G" "G" "T" "G" "C" "A" "G" "A" "G" "A" "C" "C" "A" "G" "C" "G" "A" "A"
#>  [9325] "A" "A" "A" "T" "G" "G" "A" "A" "C" "C" "G" "T" "G" "A" "T" "G" "G" "A"
#>  [9343] "T" "G" "T" "C" "A" "T" "A" "T" "C" "C" "A" "G" "A" "C" "G" "T" "G" "A"
#>  [9361] "C" "C" "A" "G" "A" "G" "A" "G" "G" "A" "A" "G" "T" "G" "G" "A" "C" "A"
#>  [9379] "G" "G" "T" "T" "G" "G" "A" "A" "C" "C" "T" "A" "T" "G" "G" "C" "T" "T"
#>  [9397] "A" "A" "A" "C" "A" "C" "C" "T" "T" "C" "A" "C" "C" "A" "A" "C" "A" "T"
#>  [9415] "G" "G" "A" "G" "G" "C" "C" "C" "A" "A" "C" "T" "A" "A" "T" "A" "A" "G"
#>  [9433] "A" "C" "A" "A" "A" "T" "G" "G" "A" "G" "T" "C" "T" "G" "A" "G" "G" "G"
#>  [9451] "A" "A" "T" "C" "T" "T" "T" "T" "C" "A" "C" "C" "C" "A" "G" "C" "G" "A"
#>  [9469] "A" "T" "T" "G" "G" "A" "A" "A" "C" "C" "C" "C" "A" "A" "A" "T" "C" "T"
#>  [9487] "A" "G" "C" "C" "G" "A" "A" "A" "G" "A" "G" "T" "C" "C" "T" "C" "G" "A"
#>  [9505] "C" "T" "G" "G" "T" "T" "G" "A" "A" "A" "A" "A" "A" "C" "A" "T" "G" "G"
#>  [9523] "C" "A" "C" "C" "G" "A" "G" "A" "G" "G" "C" "T" "G" "A" "A" "A" "A" "G"
#>  [9541] "A" "A" "T" "G" "G" "C" "A" "A" "T" "C" "A" "G" "T" "G" "G" "A" "G" "A"
#>  [9559] "T" "G" "A" "C" "T" "G" "T" "G" "T" "G" "G" "T" "G" "A" "A" "A" "C" "C"
#>  [9577] "A" "A" "T" "C" "G" "A" "T" "G" "A" "C" "A" "G" "A" "T" "T" "T" "G" "C"
#>  [9595] "A" "A" "C" "A" "G" "C" "C" "T" "T" "A" "A" "C" "A" "G" "C" "T" "T" "T"
#>  [9613] "G" "A" "A" "T" "G" "A" "C" "A" "T" "G" "G" "G" "A" "A" "A" "G" "G" "T"
#>  [9631] "A" "A" "G" "A" "A" "A" "A" "G" "A" "C" "A" "T" "A" "C" "C" "G" "C" "A"
#>  [9649] "A" "T" "G" "G" "G" "A" "A" "C" "C" "T" "T" "C" "A" "A" "A" "A" "G" "G"
#>  [9667] "A" "T" "G" "G" "A" "A" "T" "G" "A" "T" "T" "G" "G" "C" "A" "A" "C" "A"
#>  [9685] "A" "G" "T" "G" "C" "C" "T" "T" "T" "C" "T" "G" "T" "T" "C" "A" "C" "A"
#>  [9703] "C" "C" "A" "T" "T" "T" "C" "C" "A" "C" "C" "A" "G" "C" "T" "G" "A" "T"
#>  [9721] "T" "A" "T" "G" "A" "A" "G" "G" "A" "T" "G" "G" "G" "A" "G" "G" "G" "A"
#>  [9739] "G" "A" "T" "A" "G" "T" "G" "G" "T" "G" "C" "C" "A" "T" "G" "C" "C" "G"
#>  [9757] "C" "A" "A" "C" "C" "A" "A" "G" "A" "T" "G" "A" "A" "C" "T" "T" "G" "T"
#>  [9775] "A" "G" "G" "T" "A" "G" "G" "G" "C" "C" "A" "G" "A" "G" "T" "A" "T" "C"
#>  [9793] "A" "C" "A" "A" "G" "G" "C" "G" "C" "C" "G" "G" "A" "T" "G" "G" "A" "G"
#>  [9811] "C" "T" "T" "G" "A" "G" "A" "G" "A" "A" "A" "C" "T" "G" "C" "A" "T" "G"
#>  [9829] "C" "C" "T" "A" "G" "G" "C" "A" "A" "G" "T" "C" "A" "T" "A" "T" "G" "C"
#>  [9847] "A" "C" "A" "A" "A" "T" "G" "T" "G" "G" "C" "A" "G" "C" "T" "G" "A" "T"
#>  [9865] "G" "T" "A" "C" "T" "T" "C" "C" "A" "C" "A" "G" "G" "A" "G" "A" "G" "A"
#>  [9883] "C" "T" "T" "G" "A" "G" "A" "T" "T" "A" "G" "C" "G" "G" "C" "T" "A" "A"
#>  [9901] "T" "G" "C" "T" "A" "T" "C" "T" "G" "T" "T" "C" "A" "G" "C" "C" "G" "T"
#>  [9919] "T" "C" "C" "A" "G" "T" "T" "G" "A" "T" "T" "G" "G" "G" "T" "C" "C" "C"
#>  [9937] "A" "A" "C" "C" "A" "G" "C" "C" "G" "C" "A" "C" "C" "A" "C" "C" "T" "G"
#>  [9955] "G" "T" "C" "G" "A" "T" "C" "C" "A" "T" "G" "C" "C" "C" "A" "C" "C" "A"
#>  [9973] "T" "C" "A" "A" "T" "G" "G" "A" "T" "G" "A" "C" "A" "A" "C" "A" "G" "A"
#>  [9991] "A" "G" "A" "C" "A" "T" "G" "T" "T" "G" "T" "C" "A" "G" "T" "G" "T" "G"
#> [10009] "G" "A" "A" "T" "A" "G" "G" "G" "T" "T" "T" "G" "G" "A" "T" "A" "G" "A"
#> [10027] "G" "G" "A" "A" "A" "A" "C" "C" "C" "A" "T" "G" "G" "A" "T" "G" "G" "A"
#> [10045] "G" "G" "A" "C" "A" "A" "G" "A" "C" "T" "C" "A" "T" "G" "T" "G" "T" "C"
#> [10063] "C" "A" "G" "T" "T" "G" "G" "G" "A" "A" "G" "A" "C" "G" "T" "T" "C" "C"
#> [10081] "A" "T" "A" "C" "C" "T" "A" "G" "G" "A" "A" "A" "A" "A" "G" "G" "G" "A"
#> [10099] "A" "G" "A" "T" "C" "A" "A" "T" "G" "G" "T" "G" "T" "G" "G" "T" "T" "C"
#> [10117] "C" "C" "T" "A" "A" "T" "A" "G" "G" "C" "T" "T" "A" "A" "C" "A" "G" "C"
#> [10135] "A" "C" "G" "A" "G" "C" "C" "A" "C" "C" "T" "G" "G" "G" "C" "C" "A" "C"
#> [10153] "C" "A" "A" "C" "A" "T" "A" "C" "A" "A" "G" "T" "G" "G" "C" "C" "A" "T"
#> [10171] "A" "A" "A" "C" "C" "A" "A" "G" "T" "G" "A" "G" "A" "A" "G" "G" "C" "T"
#> [10189] "C" "A" "T" "T" "G" "G" "G" "A" "A" "T" "G" "A" "G" "A" "A" "T" "T" "A"
#> [10207] "T" "C" "T" "A" "G" "A" "C" "T" "T" "C" "A" "T" "G" "A" "C" "A" "T" "C"
#> [10225] "A" "A" "T" "G" "A" "A" "G" "A" "G" "A" "T" "T" "C" "A" "A" "A" "A" "A"
#> [10243] "C" "G" "A" "G" "A" "G" "T" "G" "A" "T" "C" "C" "C" "G" "A" "A" "G" "G"
#> [10261] "G" "G" "C" "A" "C" "T" "C" "T" "G" "G" "T" "A" "A" "G" "C" "C" "A" "A"
#> [10279] "C" "T" "C" "A" "T" "T" "C" "A" "C" "A" "A" "A" "A" "T" "A" "A" "A" "G"
#> [10297] "G" "A" "A" "A" "A" "T" "A" "A" "A" "A" "A" "A" "T" "C" "A" "A" "A" "C"
#> [10315] "A" "A" "G" "G" "C" "A" "A" "G" "A" "A" "G" "T" "C" "A" "G" "G" "C" "C"
#> [10333] "G" "G" "A" "T" "T" "A" "A" "G" "C" "C" "A" "T" "A" "G" "C" "A" "C" "G"
#> [10351] "G" "T" "A" "A" "G" "A" "G" "C" "T" "A" "T" "G" "C" "T" "G" "C" "C" "T"
#> [10369] "G" "T" "G" "A" "G" "C" "C" "C" "C" "G" "T" "C" "C" "A" "A" "G" "G" "A"
#> [10387] "C" "G" "T" "A" "A" "A" "A" "T" "G" "A" "A" "G" "T" "C" "A" "G" "G" "C"
#> [10405] "C" "G" "A" "A" "A" "G" "C" "C" "A" "C" "G" "G" "T" "T" "C" "G" "A" "G"
#> [10423] "C" "A" "A" "G" "C" "C" "G" "T" "G" "C" "T" "G" "C" "C" "T" "G" "T" "A"
#> [10441] "G" "C" "T" "C" "C" "A" "T" "C" "G" "T" "G" "G" "G" "G" "A" "T" "G" "T"
#> [10459] "A" "A" "A" "A" "A" "C" "C" "C" "G" "G" "G" "A" "G" "G" "C" "T" "G" "C"
#> [10477] "A" "A" "A" "C" "C" "A" "T" "G" "G" "A" "A" "G" "C" "T" "G" "T" "A" "C"
#> [10495] "G" "C" "A" "T" "G" "G" "G" "G" "T" "A" "G" "C" "A" "G" "A" "C" "T" "A"
#> [10513] "G" "T" "G" "G" "T" "T" "A" "G" "A" "G" "G" "A" "G" "A" "C" "C" "C" "C"
#> [10531] "T" "C" "C" "C" "A" "A" "G" "A" "C" "A" "C" "A" "A" "C" "G" "C" "A" "G"
#> [10549] "C" "A" "G" "C" "G" "G" "G" "G" "C" "C" "C" "A" "A" "C" "A" "C" "C" "A"
#> [10567] "G" "G" "G" "G" "A" "A" "G" "C" "T" "G" "T" "A" "C" "C" "C" "T" "G" "G"
#> [10585] "T" "G" "G" "T" "A" "A" "G" "G" "A" "C" "T" "A" "G" "A" "G" "G" "T" "T"
#> [10603] "A" "G" "A" "G" "G" "A" "G" "A" "C" "C" "C" "C" "C" "C" "G" "C" "A" "C"
#> [10621] "A" "A" "C" "A" "A" "C" "A" "A" "A" "C" "A" "G" "C" "A" "T" "A" "T" "T"
#> [10639] "G" "A" "C" "G" "C" "T" "G" "G" "G" "A" "G" "A" "G" "A" "C" "C" "A" "G"
#> [10657] "A" "G" "A" "T" "C" "C" "T" "G" "C" "T" "G" "T" "C" "T" "C" "T" "A" "C"
#> [10675] "A" "G" "C" "A" "T" "C" "A" "T" "T" "C" "C" "A" "G" "G" "C" "A" "C" "A"
#> [10693] "G" "A" "A" "C" "G" "C" "C" "A" "A" "A" "A" "A" "A" "T" "G" "G" "A" "A"
#> [10711] "T" "G" "G" "T" "G" "C" "T" "G" "T" "T" "G" "A" "A" "T" "C" "A" "A" "C"
#> [10729] "A" "G" "G" "T" "T" "C" "T"